refactor(wiki): Stage C.1 — split core/wiki_generator.py 51KB → package, all files ≤17KB
Splits the 51 KB monolithic `core/wiki_generator.py` into a mixin-based
package so every file respects CLAUDE.md rule #5 (< 20 KB). This is
the first of the Stage C module-size cleanups identified in the
2026-05-23 audit; resolves PLATFORM_READINESS Gate v0.3 dimension A
for this file (the largest of five violations).

## What changed

- `core/wiki_generator.py` (51 KB, 1218 lines, 1 class, 19 methods)
  → `core/wiki_generator/` package with the same external surface.
- Composition uses a 4-mixin facade. MRO:
  WikiIngestionMixin → WikiMergeMixin → WikiIndexOpsMixin →
  WikiFrontmatterMixin → object.
- No method signatures, attributes, or call paths changed. The split
  is purely textual.

## File sizes (post-split)

  __init__.py         2950 B   facade, re-exports, class composition
  _aliases.py         4446 B   regex/constants + synonyms loader + alias expander
  _frontmatter.py    16844 B   __init__, ID gen, search, create_entity_file
  _index_ops.py       5053 B   update_index, resolve_pending_relations, stats
  _ingestion.py      16281 B   process_document, LLM extract, safe filter
  _merge.py          11952 B   relation merge, build_entity_relations, inverse label

All six files comfortably under the 20 KB gate.

## External surface preserved

All current import paths work unchanged:

- `from core.wiki_generator import WikiGenerator` — 14 sites
- `from core.wiki_generator import _expand_alias_candidates` — scripts/migrate_aliases.py:33
- `from core.wiki_generator import _ALLOWED_EXTRACT_TYPES` — tests/test_event_ingest_emit.py:342
- `import core.wiki_generator as wg_mod` (mock patching) — 10+ sites
- `wg_mod.WIKI_DIR = tmp` monkey-patch — preserved via:
  - `core/wiki_generator/__init__.py` imports `WIKI_DIR` from `config` at package level
  - `WikiFrontmatterMixin.__init__` and `_aliases._load_synonyms` do
    `from core.wiki_generator import WIKI_DIR` lazily so the patch is
    visible at call time.

## Mixin layout (Layer → Mixin)

- L0 utilities (`_normalize_name`, `_read_frontmatter`, ID gen,
  default sensitivity, `_register_entity_id`) → _frontmatter.py
- L1 index/search (`_build_entity_id_index`, `refresh_entity_map`,
  `_find_existing_entity_id`, `find_duplicate_entities`) →
  _frontmatter.py
- L2 single-entity writer (`create_entity_file`) → _frontmatter.py
- L2.5 wiki-wide sweeps (`update_index`, `resolve_pending_relations`,
  `get_entity_statistics`) → _index_ops.py
- L3 cross-doc merge (`_merge_relations_into_existing_entity`,
  `_build_entity_relations`, `_inverse_label_for`) → _merge.py
- L4 orchestrator (`process_document_for_entities`, LLM extract,
  safe filter) → _ingestion.py

The `WIKI_DIR` monkey-patch path was the riskiest invariant — the
audit's plan agent (R1 mitigation) flagged it before any code moved.
Late-bound `from core.wiki_generator import WIKI_DIR` inside
`WikiFrontmatterMixin.__init__` keeps the test pattern working
byte-identical.

## Verification

- All 5 highest-risk monkey-patch test files pass (66/66):
  - tests/test_phase_b_ingestion_sources.py
  - tests/test_event_ingest_emit.py
  - tests/test_phase_d_modify_cascade.py
  - tests/test_phase_c_cascade.py
  - tests/test_wiki_resolve_unresolved.py
- Full pytest suite (matching CI's ignore list): **1590 passed, 824
  subtests passed**. Zero regression.
- `scripts/dogfood_check.py` — 4/4 plugin invariants still PASS.
- `ruff check core/wiki_generator/` — clean.
- Sanity import: `WikiGenerator.__mro__` shows 4 mixins in the
  documented order; `dir(WikiGenerator)` shows all 20 original
  methods (19 named + `__init__`).

## Stage C progress

- C.1 (wiki_generator.py 51 KB) — this PR ✅
- C.2 (cascade.py 24 KB) — pending
- C.3 (character_profile.py 24 KB) — pending
- C.4 (security_layer.py 23 KB) — pending (after Stage D.1 JSONL writers)
- C.5 (graph_editor.py 20 KB boundary) — pending; reassess after C.4

## Out of scope

- Multi-file pack support (intra-pack imports inside packs/) — separate
  loader concern, already documented in `core/plugins/loader.py`.
- Byte-identical STEP 7 snapshot replay — no behaviour change in this
  PR makes that gate trivially preserved; the dogfood gate (4 runtime
  invariants) is the active check.
